LayoutTests/platform/blackberry/editing/pasteboard/copy-paste-float-expected.txt

CONSOLE MESSAGE: line 18: TypeError: Result of expression 'pasteSection.firstChild.style' [undefined] is not an object.
This test is to see if text within a floating block is put onto the paste board wihtout the float style.

First LI
Second LI

---
---
